+++ src/store/aiInfra/slices/aiProvider/action.ts
@@ -19,6 +19,7 @@ import {
   UpdateAiProviderConfigParams,
   UpdateAiProviderParams,
 } from '@/types/aiProvider';
+import { getModelPropertyWithFallback } from '@/utils/getFallbackModelProperty';
 
 enum AiProviderSwrKey {
   fetchAiProviderItem = 'FETCH_AI_PROVIDER_ITEM',
@@ -203,6 +204,8 @@ export const createAiProviderSlice: StateCreator<
         onSuccess: async (data) => {
           if (!data) return;
 
+          const { LOBE_DEFAULT_MODEL_LIST } = await import('@/config/aiModels');
+
           const getModelListByType = (providerId: string, type: string) => {
             const models = data.enabledAiModels
               .filter((model) => model.providerId === providerId && model.type === type)
@@ -212,7 +215,9 @@ export const createAiProviderSlice: StateCreator<
                 displayName: model.displayName ?? '',
                 id: model.id,
                 ...(model.type === 'image' && {
-                  parameters: (model as AIImageModelCard).parameters,
+                  parameters:
+                    (model as AIImageModelCard).parameters ||
+                    getModelPropertyWithFallback(model.id, 'parameters'),
                 }),
               }));
 
@@ -231,7 +236,6 @@ export const createAiProviderSlice: StateCreator<
             children: getModelListByType(provider.id, 'image'),
             name: provider.name || provider.id,
           }));
-          const { LOBE_DEFAULT_MODEL_LIST } = await import('@/config/aiModels');
 
           set(
             {